Crate port_selector

Modules

take_up

Structs

Selector

Functions

is_free

Check whether the port is not used on TCP and UDP

is_free_tcp

Check whether the port is not used on TCP

is_free_udp

Check whether the port is not used on UDP

random_free_port

The system randomly assigns available TCP and UDP ports

random_free_tcp_port

The system randomly assigns available TCP ports

random_free_udp_port

The system randomly assigns available UDP ports

select_free_port

Gets a matching port based on the Selector parameter constraint

select_from_given_port

Check from given_port and return the first available port Return if given_port is available; Otherwise given_port += given_port until the port is available

Type Aliases

Port